Move Triangulation_structural_filtering_traits and specialize in Epic and Epec headers
